Initial libslub integration prototype
This implements a debug interface required by libslub, and add to all of the
libslub commands as callable pwndbg commands, up the form "slubxxx", whereas
libslub itself would use "sbxxx". This currently allows the two different
command sets to be used at the same time, which allows testing to ensure
both implementations work the same.

This also implements a significant number of tests for libslub functionality
in general, which will be useful once the libslub integration is complete.